mem: TwoLevel-v1.0
In last version, there was a complex logic bug in TwoLevel code. I didn't find it until i run the
MiBench-typeset workload. It also can be triggerd by the sequence-writing testing program  which
indicates that we made mistakes in the caching behavior. In order to locate the bug, I dispart the
code into several logical part and test them seperately. We take apart the code into parts: Hit, Miss, replace, writeback, and LLSC. we redirect
all LLSC request into memory directly(ignore them in cache).
In this version, I test the Missing progress, ensure that the inserting is right, and we can access
cache normally in read/write(except LLSC, they were send to memory). But, we didn't verify the
replacing, writeback.

+++ src/hum/twolevel/twolevel.hh
@@ -17,7 +17,7 @@
 struct Tag
 {
     bool dirty;
-    uint8_t position;
+    int position;
 };
 
 class TwoLevel : public ClockedObject
